requirements.txt
crispy-bootstrap5==0.7
Django==4.1.3
django-cleanup==6.0.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-db-logger==0.1.12
django-extensions==3.2.1
django-taggit==3.1.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-redis==5.2.0
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
python-slugify==7.0.0
six==1.16.0
phonenumbers==8.13.1
Pillow==9.3.0
Unidecode==1.3.6
gunicorn==20.1.0